5.0
Quick overview:
This update includes the new CTS system to save/load variables, AdvancedIndexEvent,
improvement of the BasicIndexEvent, the removal of the EntitySystem, as well as general change of directions, namely:
* From now on, the library will focus on keeping backwards compatibility as much
as possible, instead of replacing everything and changing stuff constantly,
while this approach is optimal performance wise, it is very bad for large projects
which already implemented the library on a large scope; So as of 5.0 and beyond you shouldn't
worry about not updating the library as all changes will be focued on making the
user of the library make as little changes as possible.
* Consistent naming conventions.
* Separating system into different .dll files, you will already see this with the CTS system
which is separate from the library's core WarWolfWorks.dll file and can be used
without dependency on either Unity or the WarWolfWorks library.
* Focus on advanced users: While the library will still have beginner-friendly
systems here and there from time to time, it's safe to assume that most features
from now on will be focused on advanced users.

Added:
	- CTS System: A new save system which will replace the current Streaming.Catalog system;
	Includes a much more performant way of handling variables, as well as enormous utility
	compared to it's counterpart (more on Wiki in time).
	Most importantly, it is a separate .dll file which has no dependency on Unity,
	which means that you can use the CTS system without needing Unity!
	- AdvancedIndexEvent: An advanced version of the BasicIndexEvent,
	it allows for custom scriptable objects to be implemented as part of the
	class and makes multi-function menu events much easier and faster to implement.
	There are 2 pre-made classes included:
		* AnchorsEventGraphic: Moves anchors of the given graphic based on it's focused state.
		* ColorEventGraphic: Moves the color of the given graphic based on it's focused state.
	- interface IFocusable.
	- interface IOnFocus.
	- interface IOnUnfocus.
	- Custom editor for BasicIndexEvent.

Misc:
	- Entire WarWolfWorks library is updated to use the new CTS system to
	save variables (like DefaultKeys, WarWolfWorks/Settings, etc...).
	- Changed naming convention for const variables:
		* Streaming variables will all start with SV instead of SVARV.
		* Streaming variable names will all start with SVN instead of SVARN.
		* Categories will all start with SVCN instead of SVARNCN.
		* Normal variables will all start with V instead of VARV.
		* Variable names will all start with VN instead of VARN.
	- Renamed SVARCN_UTIL_CANVAS to SV_CORECANVAS.
	- Renamed SVARNPTH_KEYS to SV_Path_DefaultKeys.
	- Renamed Path_Settings to SV_Path_Settings.
	- Renamed Path_Preferences to SV_Path_Preferences.
	- All Streaming.Catalog-related methods are now attributed with Obsolete.

Removed:
	- EntitySystem.